Add links to Minder rule types.
This change adds links to existing Minder rule types that implement
Baseline checks.

Implementation is based on what was done for Scorecard probes, but
definitions in `baseline.yaml` are a bit more complex since some
Minder rules can be configured by the end user, and that bit of code
lives in a Profile in Minder lingo.